+++ apps/backend/src/main/kotlin/hu/kirdev/foodex/dto/FoodExRequestDTO.kt
@@ -5,10 +5,10 @@ import java.time.LocalDateTime
 
 data class FoodExRequestDTO(
     @field:NotBlank
-    val userId: Long,
+    val userId: Int,
 
     @field:NotBlank
-    val cookingClubId: Long,
+    val cookingClubId: Int,
 
     @field:NotBlank
     val opening: LocalDateTime,
