+++ apps/backend/src/main/kotlin/hu/kirdev/foodex/repository/CookingClubRepository.kt
@@ -4,12 +4,12 @@ import hu.kirdev.foodex.model.CookingClubEntity
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
 
-interface CookingClubRepository : JpaRepository<CookingClubEntity, Long> {
-    fun findCookingClubEntityById(id: Long): CookingClubEntity?
+interface CookingClubRepository : JpaRepository<CookingClubEntity, Int> {
+    fun findCookingClubEntityById(id: Int): CookingClubEntity?
 
     @Query("SELECT c FROM CookingClubEntity c WHERE :leaderId IN (c.leaders)")
-    fun findAllByLeaderId(leaderId: Long): List<CookingClubEntity>
+    fun findAllByLeaderId(leaderId: Int): List<CookingClubEntity>
 
     @Query("SELECT c.leaders FROM CookingClubEntity c WHERE c.id = :clubId")
-    fun findLeadersByCookingClubId(cookingClubId: Long): List<Long>
+    fun findLeadersByCookingClubId(cookingClubId: Int): List<Int>
 }
